\docType{data}
\name{andalusia-data}
\alias{andalusia}
\alias{andalusia-data}
\title{Cities of Andalusia}
\format{\code{name}: The name of the city or relative position label.

\code{x}: The x coordinate of points.

\code{y}: The y coordinate of points.

\code{city}: If yes the point is a city in other case is a limit.}
\description{
  The 'andalusia' data frame has 12 rows and 4 columns,
  which are the geographical position of the main capital
  cities of andalusia.
}
\section{Usage}{
  data('andalusia')
}

\section{Source}{
  The data are taken from wikipedia.
}
\seealso{
  See also \code{\link{orloca-package}}.
}
\keyword{andalusia}
\keyword{data}
\keyword{datasets}

