\name{andalusia-data}
\docType{data}
\alias{andalusia}

\title{Cities of Andalusia}
\description{The 'andalusia' data frame has 12 rows and 4 columns, which
  are the geographical position of the main capital cities of andalusia.}
\usage{andalusia}
\section{Format}{
  \code{name}: The name of the city or relative position label.
  
  \code{x}: The x coordinate of points.
  
  \code{y}: The y coordinate of points.
  
  \code{city}: If yes the point is a city in other case is a limit.
}
\source{The data are from wikipedia.}
\keyword{data}
\keyword{andalusia}
\seealso{See also \code{\link{orloca-package}}.}