% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distsumgra.R
\docType{methods}
\name{zsumgra}
\alias{zsumgra}
\title{zsumgra}
\usage{
zsumgra(...)
}
\arguments{
\item{\ldots}{Parameters passed to distsumgra}
}
\description{
The function zsumgra is deprected and could be removed in next version of the package. Use \link{distsumgra} instead.
}
\keyword{deprecated}
