% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsuml2min.R
\name{zsuml2min}
\alias{zsuml2min}
\alias{zsuml2min,loca.p-method}
\title{zsuml2min at orloca package}
\usage{
zsuml2min(o, x = 0, y = 0, max.iter = 100, eps = 0.001,
  verbose = FALSE, algorithm = "weiszfeld", ...)
}
\arguments{
\item{o}{An object of loca.p class.}

\item{x}{The x coordinate of the starting point.}

\item{y}{The y coordinate of the starting point.}

\item{max.iter}{Maximum number of iterations allowed.}

\item{eps}{The module of the gradient in the stop rule.}

\item{verbose}{If TRUE the function produces detailed output.}

\item{algorithm}{The method to be use. For this version of the package, the valid values are: "gradient" or "g" for a gradient based method, "search" or "s" for local search method (this option is deprecated), "ucminf" or "u" for optimization with ucminf from ucminf package, and "weiszfeld" or "w" for the weiszfeld method or any of the valid method for optim function, now "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN". "weiszfeld" is the default value.}

\item{\ldots}{Other options for optimization algorithms.}
}
\value{
\code{zsuml2min} returns an array with the coordinates of the solution point.
}
\description{
\code{zsummin} function for the Euclidean norm (\eqn{l_2}).
This function returns the solution of the minimization problem.
Mainly for internal use.
}
\seealso{
See also \code{\link{orloca-package}}, \code{\link{zsummin}}, \code{\link{loca.p}} and \code{\link{zsum}}.
}
\keyword{classes}
\keyword{internal}
\keyword{optimize}
