% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distsummin.R
\docType{methods}
\name{zsummin}
\alias{zsummin}
\title{zsummin}
\usage{
zsummin(...)
}
\arguments{
\item{\ldots}{Parameters passed to distsummin}
}
\description{
The function zsummin is deprected and could be removed in next version of the package. Use \link{distsummin} instead.
}
\keyword{deprecated}
