% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distsum.R
\docType{methods}
\name{zsum}
\alias{zsum}
\title{zsum}
\usage{
zsum(...)
}
\arguments{
\item{\ldots}{Parameters passed to distsum}
}
\description{
The function zsum is deprected and could be removed in next version of the package. Use \link{distsum} instead.
}
\keyword{deprecated}
